2.2.Design Config & Security Setup{Adding Security}
Adding Security

Before we keep going... we need to talk security! Because right now, I can log out and go back to /easyadmin with no problem. This is totally open to the public. Fun!
How can we configure security in EasyAdminBundle? We don't! This is just a normal page... so we can use Symfony's normal security to protect it.
The easiest way is in security.yml via access_control.
Uncomment the example, use ^/easyadmin and check for ROLE_ADMIN.
That is it!
When we refresh... yep! We're bounced back to the login page. Log back in with weaverryan+1@gmail.com password iliketurtles. And we're back! We can also secure things in a more granular way... and I'll show you how later.
Now, let's start customizing the list page.