3.2.Views & entities Config{Configuring at the Entity Level}
Configuring at the Entity Level
So far, we just have a simple list of all the entity admin sections we want. That's great to get started... but not anymore! As soon as you need to configure an entity further, you need to use an expanded format. Basically, instead of - Genus, use Genus as the key and add a new line with class set to the full class name: AppBundle\Entity\Genus
Repeat this for everything else: GenusNote, SubFamily and User
This didn't change anything... but now we are, oh yes, dangerous! Oh, so much can be configured for each entity. Start simple: label: Genuses
Try that! Nice! The label is of course used in the title, but also in the navigation.