3.4.Views & entities Config{Overriding Entity View Config}
Overriding Entity View Config
In addition to title, one other key that every view has is help. First, set this below the Genus section: "Genuses are not covered by warranty! "
Notice, this is directly under the entity config, not under a specific view. Thanks to this, it will apply to all views for this entity. And... yep! It's at the top of the search page, on list and on edit.
In the spirit of EasyAdminBundle, you can override this for each view below. For list, set help to Do not feed!
Nice!
And the search view still uses the default message. If you want to turn off the help message here, you can totally do that. Under search, set help to null
Refresh! Gone!
Ok, so we're not going to talk about all of the different keys available: it's all pretty easy. The most important thing is to realize that each of the 5 views can be configured on two different levels.
Next, let's talk about actions and the different ones that we can enable.