4.1.Actions Config{Adding the show Action; Adding __toString() Methods}
We now know that there are 5 views: list, search, edit, new and banana... show. EasyAdminBundle also has an idea of "actions", which are basically the links and buttons that show up on each view. For example, on list, we have a search action, a new action and edit & delete actions. In the docs, they have a section called Actions Configuration that shows the "default" actions for each view. For example, from the "Edit" view, you have a delete button and a list link... but you do not have a link to create a new entity or eat a banana.
These actions can be customized a lot: we can add some, take some away, tweak their design and - gasp - create custom actions.

Adding the show Action
In the docs, it says that the list view does not have the show action by default. Yep, there's no little "show" link next to each item. If you want that, add it! Under the global list, add actions, then show
We only need to say show, we don't need to re-list all the actions. That's because the, actions configuration is a merge: it takes all of the original actions and adds whatever we have here.

Adding __toString() Methods
So... how would we remove an action? We'll get there! But first, if you click, "Show"... wow. Geez. Dang! The page is very broken:

  Object of class GenusNote could not be converted to string.

In a lot of places, EasyAdminBundle tries to convert your entity objects into strings... ya know, so it can print them in a list or create a drop-down menu. To get this working, we need to add __toString() methods to each entity. Let's do that real quick!