4.2.Actions Config{Adding __toString() Methods; Removing Actions}
To get actions working, we need to add __toString() methods to each entity. Let's do that real quick!
In Genus, add public function __toString() and return $this->getName(). I'm casting this into a string just to be safe - if it's null, you'll get an error. In GenusNote, do the same thing: return $this->getNote(). In GenusScientist, we can return $this->getUser().
That's an object... but we're about to add a __toString() for it too!
SubFamily, well hey! It already has a __toString(). I'll just add the string cast.
And finally, in User ... make this a bit fancier. Return $this->getFullName() or $this->getEmail(), in case the user doesn't have a first or last name in the database
Try the show page again! Nice! It renders all of the properties, including the relations, which is why it needed that __toString() method.
And because we have a SubFamily admin section, the SubFamily is a link that takes us to its show view.

Removing Actions
Speaking of SubFamily, as you can see... there's not much to it: just an id and a name. And the show view, well, it's just the id and name again. Not too interesting. In this case, I think it's overkill to have the show action for the SubFamily entity. So let's kill it!