4.4.Actions Config{Disabling Actions}
If you truly want to disable the show view, there's a special config key for that. Under the entity itself, add disabled_actions set to an array with show inside.
As soon as we do that ... the link vanishes in dramatic fashion! Let's be annoying again: I'll go forward in my browser to get back to the show page URL. Refresh! Dang! Now we get a huge error. The show view is totally gone.