6.1.Configuring the List Fields{List Field Types and Options; Changing to the boolean data type}
Navigate to the User entity section... open up the EasyAdminBundle profiler... and check out the list fields.
I want to talk a bit more about these field "data types". Check out isScientist. Its data type is set to toggle, my favorite type! Go back to the main page of the documentation and open List, Search and Show Views Configuration.

List Field Types and Options
Down the page a little, it talks about how to "Customize the Properties Appearance". This is great stuff. First, it lists the valid options for all the fields, like property, label, css_class, template - which we'll talk about later - and yes! The type, which controls that dataType config. There are a bunch of built-in types, including all of the Doctrine field types and a few special fancy ones from EasyAdminBundle, like toggle. The "toggle" type is actually super cool: it renders the field as a little switch that turns the value on and off via AJAX.

Changing to the boolean data type
Let's take control of the User fields. Below that entity, add list, then fields, with id and email.
Let's customize isScientist and set its label to Is scientist?. And as cool as the toggle field is, for the great sake of learning, change it to boolean. Then add, firstName, lastName, and avatarUri.
Try that! Ok! The isScientist field is now a "less-cool" Yes/No label. Open up the EasyAdminBundle config to see the difference. Under list... fields... isScientist, yep! dataType is now boolean... and it's using a different template to render it. More on that later.