6.2.Configuring the List Fields{Virtual Fields}
Virtual Fields
Back in the config, obviously, id, email, isScientist, firstName, lastName, avatarUri are all property names on the User entity. But... that's not required. As long as there is a "getter" method, you can invent new, "virtual" fields. Our User does not have a fullName property... but it does have a getFullName() method. So, check this out: remove firstName and lastName and replace it with fullName
Try that out! Magic!