6.3.Configuring the List Fields{The email and url Fields}
The email and url Fields

As we saw earlier, EasyAdminBundle also has a few special types. For example, expand the email property and set its type to email. While we're here, do the same for avatarUri, setting it to url. Try that! I know, it's not earth-shattering, but it is nice: the email is now a mailto link and the avatarUri is a link to open in a new tab.