6.4.Configuring the List Fields{The image Type}
The image Type
Of course, avatarUri is an image... so it would be way trendier to... ya know... actually render an image! Yea! But let's do it somewhere else: go to the GenusNote section. Then, in config.yml, under the entity's list key - add fields. Let's show id and username
One of the fields is called the userAvatarFileName, which is a simple text field that stores an image filename, like leanna.jpeg or ryan.jpeg. I want that to show up as an image thumbnail. To do that, add property: userAvatarFilename, label: User avatar and... type: image
Before we try that, also add createdAt and genus. Actually, genus is the property that points to the related Genus object, which is pretty cool. That will totally work because our Genus class has a __toString() method.
Refresh! Ok, it kinda works... there is an image tag! Yea... it's broken, but let's try to be positive! Right click to open that in a new tab. Ah, it's look for the image at http://localhost:8000/leanna.jpeg. In our simple system, those images are actually stored in a public/images directory. In a more complex app, you might store them in an uploads/ directory or - even better - up on something like S3. But no matter where you store your images, you'll need to configure this field to point to the right path.