7.1.DQL Filtering & Sorting{Configuring Sort}
What else could we possibly configure with the list view? How about sorting, or filtering list via DQL. OoooOOooo.

Configuring Sort
First, sorting... which we get for free. Already, the genuses are sorted by id, but we can click any column to sort by that. But this isn't sticky: when you come back to the genus list page, it's back to filtering by id.
Sorting by name would be a bit more awesome. And you can probably guess what the config looks like to do this. Under Genus and list, add sort: name.
This is the new default field for sorting.