7.2.DQL Filtering & Sorting{Sorting via Relations}
Sorting via Relations
Oh, but we can get fancier. Under GenusNote, what if I told you I wanted to sort by the name of the Genus it's related to? Yea, that would mean sorting across a relation. But that's totally possible: sort: ['genus.name', 'ASC']
This also controls the direction. It sorts descending by default.
Try it! Nice! This works... just don't get too confident and try to do this across multiple relationships... that's not going to work.