7.5.DQL Filtering & Sorting{DQL Filtering}
I want to filter "Horde of Genuses" to only list genuses that have a lot of species. Start by adding a list key and a new, awe-inspiring option: dql_filter. For the value, pretend that you're building a query in Doctrine. So, entity.speciedCount >= 50000
The alias will always be entity.
Try it! Ten down to... only 7 menacing genuses!