8.2.Customize all the Templates! {How to Override Templates}
But this key thing is pretty excellent, so we should use it everywhere. Copy the templates config and comment it out.
Just like with almost anything, we can also control the template globally: paste this under design
Now the key icon shows up everywhere.
Next, I want to override something bigger: the entire list template. And we'll use a different convention to do that.