9.4.Dynamically Remove the delete Action Link{ Creating the Filter Twig Exension }
Creating the Filter Twig Exension
Phew! All we need to do now is create that filter! In src/AppBundle/Twig, create a new PHP class: EasyAdminExtension. To make this a Twig extension, extend AbstractExtension.
Then, go to the Code->Generate menu - or Command+N on a Mac - and override the getFilters() method.
Here, return an array with the filter we need: new TwigFilter('filter_admin_actions', [$this, 'filterActions']).
Down below, create public function filterActions() with two arguments. First, it will be passed an $itemActions array - that's the _list_item_actions variable. And second, $item: whatever entity is being listed at that moment.
Ok, let's fill in the logic: if $item instanceof Genus && $item->getIsPublished(), then unset($itemActions['delete']). At the bottom, return $itemActions.
Phew! That should do it! This project uses the new Symfony 5 autowiring, auto-registration and autoconfigure services.yml goodness
So... we don't need to configure anything: EasyAdminExtension will automatically be registered as a service and tagged with twig.extension. In other words... it should just work.
Let's go. Refresh... and hold your breath.
Haha, it kind of worked! Delete is gone... but so is everything else.