13.1.CollectionType Field
The form system does a pretty good job guessing the correct field types... but nobody is perfect. For example, the genusScientists field is not setup correctly. Click the clipboard icon to open the form profiler.
Yep, genusScientists is currently an EntityType with multiple set to true. Thanks to EasyAdminBundle, it renders this as a cool, tag-like, auto-complete box. Fancy!
But... that's not going to work here: the GenusScientist entity has an extra field called yearsStudied:

 85 lines  src/AppBundle/Entity/GenusScientist.php
class GenusScientist
{
    /**
     * @ORM\Column(type="integer")
     * @Assert\NotBlank()
     */
    private $yearsStudied;
}
When you link a Genus and a User, we need to allow the admin to also fill in how many years the User has studied the Genus.
In the Symfony series, we did a lot of work to create a CollectionType field that used GenusScientistEmbeddedForm:

 108 lines  src/AppBundle/Form/GenusFormType.php
class GenusFormType extends AbstractType
{
    public function buildForm(FormBuilderInterface $builder, array $options)
    {
        $builder
            ->add('genusScientists', CollectionType::class, [
                'entry_type' => GenusScientistEmbeddedForm::class,
                'allow_delete' => true,
                'allow_add' => true,
                'by_reference' => false,
            ])
        ;
}

Thanks to that, in the admin, we just need to update the form to look like this.