15.2.Form Theming For a Completely Custom Field{ Rendering the Genus Scientists }
Rendering the Genus Scientists
Let's see a quick example to get the creative juices flowing! Let's create a table that lists all of the genus scientists.
Inside a tbody, we're ready to loop over the scientists! But... uh... how can I get them? What variables do I have access to right here?
Go back to the form profiler, find genusScientists and look again at the view variables. These are all the variables that we have access to from within our form theme. But because we set the field to mapped false... um... we actually don't have access to our Genus object! That's a problem. But! Because we're inside EasyAdminBundle, it gives us a special easyadmin variable... with an item key equal to our Genus! Phew!
Ok! In the table, loop: for genusScientist in easyadmin.item.genusScientists.
Add the tr and print out a few fields: genusScientist.user and genusScientist.yearsStudied.
Let's also add a fake delete link with a class and a data-url attribute. But leave it blank.
In your app, you might create a delete AJAX endpoint and use the path() function to put that URL here so you can read it in JavaScript.
Cool! To make this a bit more realistic, open custom_backend.js. Let's find those .js-delete-scientist elements and, on click, call a function. Add the normal e.preventDefault() and... an alert('to do').
The rest, is homework!
Let's try it! There it us! A nice table with a delete icon. There's more work to do, but you can totally do it! This is just normal coding: create a delete endpoint, call it via JavaScript and celebrate!
With form stuff behind us, let's turn to adding custom actions, like, a publish button.