16.1.Adding a Custom Action{ Custom Route Actions; Adding the Route Action Endpoint }
We know there are a bunch of built-in actions, like "delete" and "edit". But sometimes you need to manipulate an entity in a different way! Like, how could we add a "publish" button next to each Genus?
There are... two different ways to do that. Click into the show view for a Genus. On show, the actions show up at the bottom. Before we talk about publishing, I want to add a new button down here called "feed"... ya know... because Genuses get hungry. When we click that, it should send the user to a custom controller where we can write whatever crazy code we want.

Custom Route Actions
The first step should feel very natural. We already know how to add actions, remove actions and customize how they look. Under Genus, add a new show key and actions. Use the expanded configuration, with name: genus_feed and type: route.
There are two different custom action "types": route and action. Route is simple: it creates a new link to the genus_feed route. And you can use any of the normal action-configuring options, like label, css_class: 'btn btn-info or an icon.

Adding the Route Action Endpoint
Next, we need to actually create that route and controller. In src/Controller, open GenusController. At the top, add feedAction() with @Route("/genus/feed") and name="genus_feed" to match what we put in the config.
Notice the URL for this is just /genus/feed. It does not start with /easyadmin. And so, it's not protected by our access_control security.
That should be enough to get started. Refresh! There's our link! Click it and... good! Error! I love errors! Our action is still empty.
So here's the question: when we click feed on the Genus show page... the EasyAdminBundle must somehow pass us the id of that genus... right? Yes! It does it via query parameters... which are a bit ugly! So I'll open up my profiler and go to "Request / Response". Here are the GET parameters. We have entity and id!
Now that we know that, this will be a pretty traditional controller. I'll type-hint the Request object as an argument.
Then, fetch the entity manager and the  via $request->query->get('id'). Use that to get the $genus object: $em->getRepository(Genus::class)->find($id)
Cool! To feed the Genus, we'll re-use a feed() method from a previous tutorial. Start by creating a menu of delicious food: shrimp, clams, lobsters and... dolphin! Then choose a random food, add a flash message and call $genus->feed()
Now that all this hard work is done, I want to redirect back to the show view for this genus. Like normal, return $this->redirectToRoute(). And actually, EasyAdminBundle only has one route... called easyadmin
We tell it where to go via query parameters, like action set to show, entity set to $request->query->get('entity')... or we could just say Genus, and id set to $id.
That is it! Refresh the show page! And feed the genus. Got it! We can hit that over and over again. Hello custom action.