16.2.Adding a Custom Action{ Custom Controller Action }
Custom Controller Action
There's also another way of creating a custom action. It's a bit simpler and a bit stranger... but has one advantage: it allows you to create different implementations of the action for different entities.
Let's try it! In config.yml, add another action. This time, set the name to changePublishedStatus with a css_class set to btn.
Let's do as little work as possible! So...refresh! We have a button! Click it! Bah! Big error! But, it explains how the feature works:

  The "changePublishedStatusAction()" method does not exist in the EasyCorp\Bundle\EasyAdminBundle\Controller\EasyAdminController class

Eureka! All we need to do is create that method... then celebrate!