16.4.Adding a Custom Action{ Pointing to our AdminController Classs }
Pointing to our AdminController Class

We totally got the exact same error! What?
This is because we haven't told Symfony to use our AdminController yet: it's still using the one from the bundle. The fix is actually in config/routes/easy_admin.yaml.
This tells Symfony to import the annotation routes from the bundle's AdminController class... which means that class is used when we go to those routes. Change this to import routes from ../../src/Controller/EasyAdmin/AdminController.php instead.
It will still read the same route annotations from the base class, because we're extending it. But now, it will use our class when that route is matched.
That should be all we need. Try it. Boom! Genus published. Do it again! Genus unpublished! The power... it's intoxicating...
Next! We're going to go rogue... and start adding our own custom hooks... like right before or after an entity is inserted or updated.