17.1.Override Controllers{ The protected AdminController Methods; Per-Entity Override Methods; One Controller per Entity }
When we submit a form, obviously, EasyAdminBundle is taking care of everything: handling validation, saving and adding a flash message. That's the best thing ever! Until... I need to hook into that process... then suddenly, it's the worst thing ever! What if I need to do some custom processing right before an entity is created or updated?
There are 2 main ways to hook into EasyAdminBundle...and I want you to know both. Open the User entity. It has an updatedAt field.
To set this, we could use Doctrine lifecycle callbacks or a Doctrine event subscriber.
But, I want to see if we can set this instead, by hooking into EasyAdminBundle. In other words, when the user submits the form for an update, we need to run some code.

The protected AdminController Methods
The first way to do this is by adding some code to the controller. Check this out: open up the base AdminController from the bundle and search for protected function. Woh... There are a ton of methods that we can override, beyond just the actions. Like, createNewEntity(), persistEntity() and opdateEntity().
If we override updateEntity() in our controller, that will be called instead of the original one from EasyAdmin. There are a few other cool things that you can override too.

Per-Entity Override Methods
Ok, easy! Just add preUpdateEntity() to our AdminController, right? Yep... but we can do better! If we override updateEntity(), it will be called whenever any entity is updated. But we really only want it to be called for the User entity.
Once again, EasyAdminBundle has our back. Inside the base controller, search for updateEntityProperty. Check this out: right before saving, it calls some executeDynamicMethod function and passes it update, a weird <EntityName> string, then Entity.
Actually, the bundle does this type of thing all over the place. Like above, when it calls createEditForm(). Whenever you see this, it means that bundle will first look for an entity-specific version of the method - like updateUserEntity() - and call it. If that doesn't exist, it will call the normal updateEntity().
This is huge: it means that each entity class can have its own set of hook methods in our AdminController!

One Controller per Entity
And now that I've told you that... we're going to do something completely different. Instead of having one controller - AdminController - full of entity-specific hook methods like preUpdateUserEntity or createGenusEditForm - I prefer to create a custom controller class for each entity.
Try this: in the EasyAdmin directory, copy AdminController and rename it to UserController. Then, remove the function. Use the "Code"->"Generate menu" - or Command+N on a Mac - to override the updateEntity() method. And don't forget to update your class name to UserController.
We're going to configure things so that this UserController is used only for the User admin section. And that means we can safely assume that the $entity argument will always be a User object.
And that makes life easy: $entity->setUpdatedAt(new \DateTime())
But how does EasyAdminBundle know to use this controller only for the User entity? That happens in config.yml. Down at the bottom, under User, add controller: App\Controller\EasyAdmin\UserController
And just like that! We have one controller that's used for just our User.
Try it out! Let's go find a user... how about ID 20. Right now, its updateAt is null. Edit it... make some changes... and save! Go back to show and... we got it!