17.2.Override Controllers{ Organizing into a Base AdminContorller }
Organizing into a Base AdminContorller
This little trick unlocks a lot of hook points. But if you look at AdminController, it's a little messy. Because, changePublishedStatusAction() is only meant to be used for the Genus class.
But technically, this controller is being used by all entities, except User.
So let's copy AdminController and make a new GenusController! Empty AdminController completely.
Then, make sure you rename the new controller class to GenusController.
But before we set this up in config, change the extends to extends AdminController, and remove the now-unused use statement.
Repeat that in UserController.
Yep, now all of our sections share a common base AdminController class. And even though it's empty now, this could be really handy later if we ever need to add a hook that affects everything.
Love it! UserController has only the stuff it needs, GenusController holds only things that relate to Genus, and if we need to override something for all entities, we can do that inside AdminController.
Don't forget to go back to your config to tell the bundle about the GenusController. All the way on top, set the Genus controller to App\Controller\EasyAdmin\GenusController.
Now we're setup to do some really, really cool stuff.