19.1.Conditional Actions{ Preventing Action Access by Role }
Ok, new challenge! I only want this edit button to be visible and accessible if the user has ROLE_SUPERADMIN. This turns out to be a bit complicated... in part because there are two sides to it.
First, we need truly block access to that action... so that a clever user can't just hack the URL and start editing! And second, we need to actually hide the link... so that our less-than-super-admin users don't get confused.

Preventing Action Access by Role
First, let's lock down the actual controller action. How? Now we know two ways: by overriding the editAction() in UserController and adding a security check or by adding a PRE_EDIT event listener. Let's use events!
Subscribe to a second event: EasyAdminEvents::PRE_EDIT set to onPreEdit.
Once again, I'll hit Alt+Enter as a shortcut to create that method for me.
And just like before... we don't really know what the  looks like. So, dump it!
Now, as soon as I hit edit... we see the dump! Check this out: this time, the subject property is actually an array. But, it has a class key set to the User class. We can use that to make sure we only run our code when we're editing a user.