19.2.Conditional Actions{ Preventing Action Access by Role }
In other words, $config = $event->getSubject() and if $config['class'] is equal to our User class, then we want to check security.
Let's call a new method... that we'll create in a moment: $this->denyAccessUnlessSuperAdmin().
At the bottom, add that: private function denyAccessUnlessSuperAdmin().
Now... we just need to check to see if the current user has ROLE_SUPERADMIN. How? Via the "authorization checker" service. To get it, type-hint a new argument with AuthorizationCheckerInterface. Hit Alt+Enter to create and set that property.
Then, back down below, if (! $this->authorizationChecker->isGranted('ROLE_SUPERADMIN'), then throw a new AccessDeniedException(). Make sure you use the class from the Security component. Oh, and don't forget the new!
See, normally, in a controller, we call $this->denyAccessUnlessGranted(). When we do that, this is actually the exception that is thrown behind the scenes. In other words, we're really doing the same thing that we normally do in a controller.
And... we're done! The service is set to be autowired, so Symfony will know to pass us the authorization checker automatically. Refresh!
Great news! Access denied! Woohoo! I've never been so happy to get kicked out of something. Our user does not have ROLE_SUPERADMIN - just ROLE_ADMIN and ROLE_USER.