19.4.Conditional Actions{ Hiding the Edit Button }
Hiding the Edit Button

Time for step 2! On the list page, we need to hide the edit link, unless I have the role. This is trickier: there's no official hook inside of EasyAdminBundle to conditionally hide or show actions. But don't worry! Earlier, we overrode the list template so that we could control exactly what actions are displayed. Our new filter_admin_actions() filter lives in EasyAdminExtension.
And we added logic there to hide the delete action for any published genuses.
In other words, we added our own hook to control which actions are displayed. We rock!
To hide the edit action, we'll need the authorization checker again. No problem! Make EasyAdminExtension to implements ServiceSubscriberInterface. Add getSubscribedServices required static method and return AuthorizationCheckerInterface service in it. Also add public function __construct with ContainerInterface as argument. Press Alt+Enter and initialize this field.
Then, down below, we'll add some familiar code: if  instanceof User and ! $this->container->get(AuthorizationCheckerInterface::class)->isGranted('ROLE_SUPERADMIN'), then unset the edit action.
Phew! It's not the easiest thing ever... EasyAdminBundle... but it does get the job done!
Except for one... minor problem... there is also an edit button on the show page. Oh no! It looks like we need to repeat all of this for the show template!