20.1.CSV Export{Defining the Custom Action; Adding the Custom Link (Conditionally)} When we started this tutorial, we kept getting the same question:
  Ryan, would you rather have rollerblades for feet or chopsticks for hands for the rest of your life?

I don't know the answer, but Eddie from Southern Rail can definitely answer this.

We also got this question:

  How can I add a "CSV Export" button to my admin list pages?

And now we know why you asked this question: it's tricky! I want to add a button on top that says Export. But, when you add a custom action to the list page... those create links next to each item, not on top. That's not what we want!

Defining the Custom Action
So let's see if we can figure this out. First, in config.yml, under the global list, we are going to add a new action, called export.
Now, if we refresh... not surprisingly, this adds an "export" link next to every item. And if we click it, it tries to execute a new exportAction() method.
So this is a bit weird: we do not want this new link on each row - we'll fix that in a few minutes. But we do need the new export action. Why? Because as soon as we add this, it's now legal to create a link that executes exportAction(). And that means that we could manually add this link somewhere else... like on top of the list page.

Adding the Custom Link (Conditionally)
Open up our custom list.html.twig.
I'll also hold Command and click to open the parent list.html.twig from the bundle. If you scroll down a little bit, you'll find a block called global_actions. Ah, it looks like it's rendering the search field. The global_actions block represents this area on top.
In other words, if we want to add a new link here, global_actions is the place to do it! Copy that block name and override it inside of our template: global_actions and endlock.
Inside, we'll add the Export button.
But wait! I have an idea. What if we only want to add the export button to some entities? Sure, I added the export action in the global section... but we could still remove it from any other entity by saying -export. Basically, I want this button to be smart: I only want to show it if the export action is enabled for this entity.
How can we figure that out? In the parent template, you'll find a really cool if statement that checks to see if an action is enabled. Steal it!
In our case, change search to export. At this point, we can do whatever we want. So, very simply, let's add a new link that points to the export action. Add a button-action div for styling.
Then, inside, a link with btn btn-primary and an href. How can we point to the exportAction()? Remember, the bundle only has one route: easyadmin. For the parameters, use a special variable called _request_parameters. This is something that EasyAdminBundle gives us, and it contains all of the query parameters. You'll see why that's cool in a minute.
But the most important thing is to add another query parameter called action set to export.
Oh boy, that's ugly. But, it works great: it generates a route to easyadmin where action is set to export and all the existing query parameters are maintained.
Phew! Inside, add a download icon and say "Export".
Try it! Woh! We have an export button... but nothing else.