20.2.CSV Export{Adding the Custom Action; Adding the CSV Export Logic; Using DI in a Fake Action}
Adding the Custom Action
I love to forget the parent() call. Try it again. Beautiful!
When I click export, it of course looks for exportAction in our controller... in this case, GenusController.
Remember: we're not going to support this export action for all of our entities. And to make this error clearer, open AdminController - our base controller - and create a public function exportAction() that simply throws a new RuntimeException: "Action for exporting an entity is not defined".
If we configure everything correctly, and implement this method for all entities that need it, we should never see this error. But... just in case.
Now, to the real work. To add an export for genus, we have two options. First, in AdminController, we could create a public function exportGenusAction(). Remember, whenever EasyAdminBundle calls any of our actions - even custom actions - it always looks for that specially named method: export<EntityName>Action(). Or, we can be a bit more organized, and create a custom controller for each entity. That's what we've done already. So, in GenusController, add public function exportAction().

Adding the CSV Export Logic
To save time, we've already done most of the work for the CSV export. If you downloaded the starting code, in the Service directory, you should have a CsvExporter class.
Basically, we pass it a QueryBuilder, an array of column information, or the entity's class name - which is mapped to an array of column info thanks to this special function, and the filename we want. Then, it creates the CSV and returns it as a StreamedResponse. So all we need to do is call this method and return it from our controller!
I'll paste a little bit of code in the action to get us started.
When we created the export link, we kept the existing query parameters. That means we should have a sortDirection parameter... which is a nice way of making the export order match the list order.
To create the query builder, we can actually use a protected function on the base class called createListQueryBuilder().
Pass this the entity class, either Genus::class or $this->entity['class']... in case you want to make this method reusable across multiple entities.
Next, pass the sort direction and then the sort field: $this->request->query->get('sortField'). Finally, pass in the dql_filter option: $this->entity['list']['dql_filter'].
This is kind of cool. We're using the entity configuration array - which is always full of goodies - to actually read the list key and the dql_filter key below it. If we have a DQL filter on this entity, the CSV export will know about it!
Ok, finally, we're ready to use the CsvExporter class.

Using DI in a Fake Action
But... remember: this is not a real action. I mean, it's not called by the normal core, Symfony controller system. Nope, it's called by EasyAdminBundle... and none of the normal controller argument tricks work. You can't type-hint the Request or any services.
Because of this, we're going to use classic dependency injection. Add a __construct() function and type-hint the CsvExporter class. I'll press Alt+Enter to create a property and set it.
Back down below, just return $this->csvExporter->getResponseFromQueryBuilder() and pass it the $queryBuilder, Genus::class, and genuses.csv - the filename.
Deep breath... refresh! It downloaded! Ha! In my terminal. I'll:
cat ~/Downloads/genuses.csv
There it is!
There's just one last problem: on the list page... we still have those weird export links on each row. That's technically fine... but it's super confusing. The only reason we added this export action was so that it would be a valid action to call:

config/packages/easy_admin.yml
easy_admin:
    list:
        actions: ['show', 'export']
Unfortunately, this also gave us those links!